requirements.txt
absl-py==0.13.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.1.0
altgraph==0.17.2
amqp==5.1.1
aniso8601==9.0.1
anyio==4.3.0
apikey==0.2.4
appdirs==1.4.4
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asn1crypto==1.5.1
astor==0.8.1
astroid==2.2.5
asttokens==2.4.1
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.4
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==24.2.0
audioread==3.0.1
autopep8==1.4.4
babel==2.16.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bleach==3.1.5
blinker==1.4
blis==0.7.9
boto3==1.24.89
botocore==1.27.89
cachetools==4.2.0
catalogue==2.0.8
catboost==1.2.7
celery==5.2.7
certifi==2022.9.24
cffi==1.14.2
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
confection==0.0.4
cryptography==39.0.2
cssselect2==0.7.0
cycler==0.11.0
cymem==2.0.7
dataframe-image==0.1.1
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
distro==1.9.0
Django==4.2.16
django-allauth==0.39.1
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-crispy-forms==1.13.0
django-environ==0.11.2
django-filter==21.1
django-hosts==6.0
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-messages==0.6.0
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-phonenumber-field==7.1.0
django-rest-auth==0.9.5
django-shopping-cart==0.1
django-utils-six==2.0
django-widget-tweaks==1.4.12
djangorestframework==3.15.2
dnspython==2.1.0
docx2txt==0.8
email-validator==1.1.2
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.0.2
executing==2.1.0
Faker==30.8.1
fastjsonschema==2.20.0
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-rabbitmq==0.0.9
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flatbuffers==1.12
fonttools==4.29.1
fqdn==1.5.1
frozenlist==1.3.0
future==1.0.0
gast==0.4.0
gevent==22.10.2
ghs==0.1.3
gitdb==4.0.5
GitPython==3.1.12
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.62.0
graphviz==0.19.1
greenlet==2.0.1
grequests==0.6.0
grpcio==1.34.1
gunicorn==19.9.0
h11==0.14.0
h5py==3.1.0
halo==0.0.31
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperopt==0.2.7
idna==2.8
importlib-metadata==7.0.1
importlib_resources==6.4.5
iniconfig==1.1.1
instagram-explore==0.3.0
ipykernel==6.29.5
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==7.6.3
isoduration==20.11.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.0.0
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab-widgets==1.0.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kombu==5.2.4
langcodes==3.3.0
lazy-object-proxy==1.4.1
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.41.1
log-symbols==0.0.14
lxml==4.9.2
Markdown==3.3.4
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.7
mccabe==0.6.1
mistune==3.0.2
msgpack==1.1.0
multidict==6.0.2
murmurhash==1.0.9
mysql-connector-python==8.0.22
mysqlclient==2.0.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netprotocols==0.5.0
networkx==3.1
nltk==3.8.1
notebook==6.1.3
notebook_shim==0.2.4
numba==0.58.1
numpy==1.24.4
oauthlib==3.0.1
openai==1.25.1
opencv-python==4.5.1.48
openpyxl==3.0.7
opt-einsum==3.3.0
oscrypto==1.3.0
outcome==1.2.0
overrides==7.7.0
packaging==24.1
pandas==1.2.1
pandocfilters==1.4.2
parso==0.7.1
pathy==0.10.1
pdfminer.six==20221105
pefile==2024.8.26
pep8==1.7.1
pickleshare==0.7.5
pika==1.3.0
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.3
plotly==5.24.1
pluggy==1.0.0
pooch==1.8.2
preshed==3.0.8
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.1
psutil==6.0.0
psycopg2==2.9.6
pure_eval==0.2.3
py==1.10.0
py4j==0.10.9.7
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.17
pydantic==1.10.6
pydeck==0.5.0
Pygments==2.6.1
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pylint==2.3.1
pymongo==4.2.0
pyparsing==2.4.7
pypdf==3.6.0
pyperclip==1.8.2
PyQRCode==1.2.1
pyresparser==1.0.6
pyrsistent==0.16.0
PySocks==1.7.1
pytest==6.2.5
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python3-openid==3.1.0
pytz==2022.4
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.13
PyYAML==6.0
pyzmq==26.2.0
qrcode==7.3.1
redis==4.5.1
referencing==0.35.1
regex==2022.10.31
reportlab==3.6.12
requests==2.32.3
requests-oauthlib==1.2.0
retry-requests==1.0.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
rsa==4.7.2
s3transfer==0.6.0
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.2
selenium==4.6.0
Send2Trash==1.8.3
six==1.12.0
sklearn==0.0
smart-open==6.3.0
smmap==3.0.5
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.1
soxr==0.3.7
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spinners==0.0.24
SQLAlchemy==1.4.9
sqlparse==0.5.1
srsly==2.4.6
stack-data==0.6.3
streamlit==0.75.0
svglib==1.5.1
tenacity==9.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==8.1.9
threadpoolctl==2.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.1
torch==1.10.2
tornado==6.4.1
tqdm==4.62.3
traitlets==5.14.3
trio==0.22.0
trio-websocket==0.9.2
typed-ast==1.4.0
typer==0.7.0
types-python-dateutil==2.9.0.20240906
typing_extensions==4.11.0
tzdata==2022.7
tzlocal==2.1
uri-template==1.3.0
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.12
uuid==1.30
validators==0.18.2
vine==5.0.0
virtualenv==20.0.31
wasabi==1.1.1
watchdog==1.0.2
wcwidth==0.2.5
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==1.0.1
wfdb==4.0.0
whitenoise==6.7.0
widgetsnbextension==3.5.1
wrapt==1.11.2
wsproto==1.2.0
WTForms==2.3.3
xhtml2pdf==0.2.9
xlrd==2.0.1
yarl==1.7.2
zipp==3.17.0
zope.event==4.5.0
zope.interface==5.5.1
